+++ src/app.ts
@@ -4,6 +4,7 @@
  */
 
 import express, { Express } from 'express';
+import cors, { CorsOptions } from 'cors';
 import {
     cleanAuthHeader,
     getRequestInformation,
@@ -17,16 +18,22 @@ import MetadataRoute from './router/routes/metadataRoute';
 import ResourceHandler from './router/handlers/resourceHandler';
 import RootRoute from './router/routes/rootRoute';
 import { applicationErrorMapper, httpErrorHandler, unknownErrorHandler } from './router/routes/errorHandling';
+import ExportRoute from './router/routes/exportRoute';
 
 const configVersionSupported: ConfigVersion = 1;
 
-export function generateServerlessRouter(fhirConfig: FhirConfig, supportedGenericResources: string[]): Express {
+export function generateServerlessRouter(
+    fhirConfig: FhirConfig,
+    supportedGenericResources: string[],
+    corsOptions?: CorsOptions,
+): Express {
     if (configVersionSupported !== fhirConfig.configVersion) {
         throw new Error(`This router does not support ${fhirConfig.configVersion} version`);
     }
     const configHandler: ConfigHandler = new ConfigHandler(fhirConfig, supportedGenericResources);
     const { fhirVersion, genericResource } = fhirConfig.profile;
     const serverUrl: string = fhirConfig.server.url;
+    let hasCORSEnabled: boolean = false;
     const app = express();
     app.use(express.urlencoded({ extended: true }));
     app.use(
@@ -36,22 +43,42 @@ export function generateServerlessRouter(fhirConfig: FhirConfig, supportedGeneri
             limit: '6mb',
         }),
     );
+    // Add cors handler before auth to allow pre-flight requests without auth.
+    if (corsOptions) {
+        app.use(cors(corsOptions));
+        hasCORSEnabled = true;
+    }
 
     // AuthZ
     app.use(async (req: express.Request, res: express.Response, next: express.NextFunction) => {
-        const requestInformation = getRequestInformation(req.method, req.path);
-        // Clean auth header (remove 'Bearer ')
-        req.headers.authorization = cleanAuthHeader(req.headers.authorization);
-        await fhirConfig.auth.authorization.isAuthorized({
-            ...requestInformation,
-            accessToken: req.headers.authorization,
-        });
-        next();
+        try {
+            const requestInformation = getRequestInformation(req.method, req.path);
+            // Clean auth header (remove 'Bearer ')
+            req.headers.authorization = cleanAuthHeader(req.headers.authorization);
+            await fhirConfig.auth.authorization.isAuthorized({
+                ...requestInformation,
+                accessToken: req.headers.authorization,
+            });
+            res.locals.requesterUserId = fhirConfig.auth.authorization.getRequesterUserId(accessToken);
+            next();
+        } catch (e) {
+            next(e);
+        }
     });
 
     // Metadata
-    const metadataRoute: MetadataRoute = new MetadataRoute(fhirVersion, configHandler);
+    const metadataRoute: MetadataRoute = new MetadataRoute(fhirVersion, configHandler, hasCORSEnabled);
     app.use('/metadata', metadataRoute.router);
+
+    // Export
+    if (fhirConfig.profile.bulkDataAccess) {
+        const exportRoute = new ExportRoute(
+            serverUrl,
+            fhirConfig.profile.bulkDataAccess,
+            fhirConfig.auth.authorization,
+        );
+        app.use('/', exportRoute.router);
+    }
 
     // Special Resources
     if (fhirConfig.profile.resources) {