+++ src/router/metadata/metadataHandler.test.ts
@@ -3,6 +3,7 @@
  *  SPDX-License-Identifier: Apache-2.0
  */
 
+import { clone, stubs } from 'fhir-works-on-aws-interface';
 import MetadataHandler from './metadataHandler';
 import { makeOperation } from './cap.rest.resource.template';
 import r4FhirConfigGeneric from '../../../sampleData/r4FhirConfigGeneric';
@@ -321,7 +322,7 @@ describe('ERROR: test cases', () => {
 
 test('STU3: FHIR Config V3 with 2 exclusions and search', async () => {
     const configHandler: ConfigHandler = new ConfigHandler(stu3FhirConfigWithExclusions, SUPPORTED_STU3_RESOURCES);
-    const metadataHandler: MetadataHandler = new MetadataHandler(configHandler);
+    const metadataHandler: MetadataHandler = new MetadataHandler(configHandler, true);
     const response = await metadataHandler.capabilities({ fhirVersion: '3.0.1', mode: 'full' });
     const { genericResource } = stu3FhirConfigWithExclusions.profile;
     const excludedResources = genericResource ? genericResource.excludedSTU3Resources || [] : [];
@@ -335,6 +336,7 @@ test('STU3: FHIR Config V3 with 2 exclusions and search', async () => {
     expect(response.resource.rest[0].resource.length).toEqual(
         SUPPORTED_STU3_RESOURCES.length - excludedResources.length,
     );
+    expect(response.resource.rest[0].security.cors).toBeTruthy();
     // see if just READ is chosen for generic
     let isExcludedResourceFound = false;
     response.resource.rest[0].resource.forEach((resource: any) => {
@@ -373,6 +375,7 @@ test('R4: FHIR Config V4 without search', async () => {
     expect(response.resource.fhirVersion).toEqual('4.0.1');
     expect(response.resource.rest.length).toEqual(1);
     expect(response.resource.rest[0].resource.length).toEqual(SUPPORTED_R4_RESOURCES.length);
+    expect(response.resource.rest[0].security.cors).toBeFalsy();
     // see if the four CRUD + vRead operations are chosen
     const expectedResourceSubset = {
         interaction: makeOperation(['create', 'read', 'update', 'delete', 'vread', 'history-instance']),
@@ -397,6 +400,7 @@ test('R4: FHIR Config V4 with 3 exclusions and AllergyIntollerance special', asy
     expect(response.resource.fhirVersion).toEqual('4.0.1');
     expect(response.resource.rest.length).toEqual(1);
     expect(response.resource.rest[0].resource.length).toEqual(SUPPORTED_R4_RESOURCES.length - excludedResources.length);
+    expect(response.resource.rest[0].security.cors).toBeFalsy();
     // see if just READ is chosen for generic
     let isExclusionFound = false;
     response.resource.rest[0].resource.forEach((resource: any) => {
@@ -440,6 +444,7 @@ test('R4: FHIR Config V4 no generic set-up & mix of STU3 & R4', async () => {
     expect(response.resource.fhirVersion).toEqual('4.0.1');
     expect(response.resource.rest.length).toEqual(1);
     expect(response.resource.rest[0].resource.length).toEqual(3);
+    expect(response.resource.rest[0].security.cors).toBeFalsy();
     // see if just READ is chosen for generic
     let isSTU3ResourceFound = false;
     response.resource.rest[0].resource.forEach((resource: any) => {
@@ -465,4 +470,32 @@ test('R4: FHIR Config V4 no generic set-up & mix of STU3 & R4', async () => {
     expect(r4Validator.validate('CapabilityStatement', response.resource)).toEqual({
         message: 'Success',
     });
+});
+test('R4: FHIR Config V4 with bulkDataAccess', async () => {
+    const r4ConfigWithBulkDataAccess = clone(r4FhirConfigGeneric);
+    r4ConfigWithBulkDataAccess.profile.bulkDataAccess = stubs.bulkDataAccess;
+    const configHandler: ConfigHandler = new ConfigHandler(r4ConfigWithBulkDataAccess, SUPPORTED_R4_RESOURCES);
+    const metadataHandler: MetadataHandler = new MetadataHandler(configHandler);
+    const response = await metadataHandler.capabilities({ fhirVersion: '4.0.1', mode: 'full' });
+
+    expect(response.resource.rest[0].operation).toEqual([
+        {
+            name: 'export',
+            definition:
+                'This FHIR Operation initiates the asynchronous generation of data to which the client is authorized. Currently only system level export is supported. For more information please refer here: http://hl7.org/fhir/uv/bulkdata/export/index.html#bulk-data-kick-off-request',
+        },
+        {
+            name: 'export-poll-status',
+            definition:
+                'After a bulk data request has been started, the client MAY poll the status URL provided in the Content-Location header. For more details please refer here: http://hl7.org/fhir/uv/bulkdata/export/index.html#bulk-data-status-request',
+        },
+    ]);
+});
+
+test('R4: FHIR Config V4 without bulkDataAccess', async () => {
+    const configHandler: ConfigHandler = new ConfigHandler(r4FhirConfigGeneric, SUPPORTED_R4_RESOURCES);
+    const metadataHandler: MetadataHandler = new MetadataHandler(configHandler);
+    const response = await metadataHandler.capabilities({ fhirVersion: '4.0.1', mode: 'full' });
+
+    expect(response.resource.rest[0].operation).toBeUndefined();
 });