+++ src/router/metadata/metadataHandler.ts
@@ -14,8 +14,11 @@ import ConfigHandler from '../../configHandler';
 export default class MetadataHandler implements Capabilities {
     configHandler: ConfigHandler;
 
-    constructor(handler: ConfigHandler) {
+    readonly hasCORSEnabled: boolean;
+
+    constructor(handler: ConfigHandler, hasCORSEnabled: boolean = false) {
         this.configHandler = handler;
+        this.hasCORSEnabled = hasCORSEnabled;
     }
 
     private generateResources(fhirVersion: FhirVersion) {
@@ -47,8 +50,8 @@ export default class MetadataHandler implements Capabilities {
         }
 
         const generatedResources = this.generateResources(request.fhirVersion);
-        const security = makeSecurity(auth);
-        const rest = makeRest(generatedResources, security, profile.systemOperations);
+        const security = makeSecurity(auth, this.hasCORSEnabled);
+        const rest = makeRest(generatedResources, security, profile.systemOperations, !!profile.bulkDataAccess);
         const capStatement = makeStatement(rest, orgName, server.url, request.fhirVersion);
 
         return {